+++ web-public/src/index.js
@@ -379,6 +379,37 @@ class WebMscore {
     }
 
     /**
+     * Read synthesis output params (master + tracks) for the current sequence.
+     * @returns {Promise<import('../schemas').AudioOutputParamsSnapshot>}
+     */
+    async getAudioOutputParams() {
+        const dataptr = Module.ccall('getAudioOutputParams',
+            'number',
+            ['number', 'number'],
+            [this.scoreptr, this.excerptId]
+        )
+        return JSON.parse(WasmRes.readText(dataptr))
+    }
+
+    /**
+     * Patch synthesis output params.
+     *
+     * Writable fields in `master` or each `tracks[i]` object will be applied.
+     * @param {import('../schemas').AudioOutputParamsPatchRequest} patch
+     * @returns {Promise<void>}
+     */
+    async setAudioOutputParams(patch) {
+        const payloadPtr = getStrPtr(JSON.stringify(patch))
+        const dataptr = Module.ccall('setAudioOutputParams',
+            'number',
+            ['number', 'number', 'number'],
+            [this.scoreptr, payloadPtr, this.excerptId]
+        )
+        freePtr(payloadPtr)
+        WasmRes.readData(dataptr)
+    }
+
+    /**
      * Synthesize audio frames
      * 
      * `synthAudio` is single instance, i.e. you can't have multiple iterators. If you call `synthAudio` multiple times, it will reset the time offset of all iterators the function returned.